+++ pkg/util/int/int_util.go
@@ -1,0 +1,10 @@
+package int
+
+// Min returns the minimum of 2 integer arguments.
+func Min(a, b int) int {
+	if a < b {
+		return a
+	}
+
+	return b
+}