+++ pkg/util/kubedns.go
@@ -10,12 +10,15 @@ import (
 // GetDnsForStatefulSet returns hostnames and names of pods in stateful set "set". This is a preferred way of getting hostnames
 // it must be always used if it's possible to read the statefulset from Kubernetes
 func GetDnsForStatefulSet(set appsv1.StatefulSet, clusterName string) ([]string, []string) {
-	return GetDNSNames(set.Name, set.Spec.ServiceName, set.Namespace, clusterName, int(*set.Spec.Replicas))
+	return GetDnsForStatefulSetReplicasSpecified(set, clusterName, 0)
 }
 
 // GetDnsForStatefulSetReplicasSpecified is similar to GetDnsForStatefulSet but expects the number of replicas to be specified
 // (important for scale-down operations to support hostnames for old statefulset)
 func GetDnsForStatefulSetReplicasSpecified(set appsv1.StatefulSet, clusterName string, replicas int) ([]string, []string) {
+	if replicas == 0 {
+		replicas = int(*set.Spec.Replicas)
+	}
 	return GetDNSNames(set.Name, set.Spec.ServiceName, set.Namespace, clusterName, replicas)
 }