+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 * Bug fixes
   * Fixes an issue which made it not possible do have multiple ops-manager resources with the same name in different namespaces
   * Fixes an issue which made new MongoDB resources created with `spec.backup.mode=disabled` fail
+  * Fixes an issue which made a Replica Set go to Fail state if, at the same time, the amount of members of a Replica Set are increased and TLS is disabled.
     
 ## MongoDBOpsManager Resource
 * Known issues